Added Flexbox for layout of the main image view
- Fixed the navbar size

- Moved the thumbnail view to top of the screen
I was running into layout issues with having a position:absolute element
taking up the bottom 150px of screen. The thumbnail bar would keep
overlapping the image being displayed. Fix for now is to move up.

- Used flexbox for layout of the app
This seems to contain the image and re-scale it dynamically to the space
at hand. Centering is a bit off and is noted in the CSS Style.